backend/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# HTTP client (pinned for FastAPI TestClient compatibility)
httpx==0.27.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.1.1

# Task Queue
celery==5.3.4
redis==5.0.0
kombu==5.3.4


# Video Processing
moviepy==1.0.3
imageio==2.33.1
imageio-ffmpeg==0.4.9
numpy>=1.26.0
Pillow==10.1.0
pydub>=0.25,<0.26

# Text-to-Speech
elevenlabs==0.2.26
requests==2.31.0

# Cloud Storage
boto3==1.34.2

# Billing
stripe>=8

# Image Generation
openai==1.13.3

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk==1.40.6
prometheus-client>=0.20.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
